admin/pyproject.toml
[project]
name = "chatroom-admin"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "William Chen", email = "william_w_chen@trendmicro.com"},
]
dependencies = [
    "py-spring-core>=0.0.6",
    "py-spring-model @ git+https://github.com/PythonSpring/pyspring-model.git@c18753a0c0cf8632f4f0232b1bf5273f44d0ca9c",
    "py-spring-admin @ git+https://github.com/PythonSpring/pyspring-admin.git@2e7b39d947096071436016a0e50694ccdd36302b",
    "kafka-python @ git+https://github.com/dpkp/kafka-python.git",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.7.0",
    "isort>=5.13.2",
]
